requirements.txt
# pip install -U -r requirements.txt
numpy~=1.19.4
opencv-python~=4.4.0.46
torch >= 1.3
matplotlib~=3.3.3
pycocotools
tqdm~=4.52.0
Pillow~=8.0.1

# Tensorboard pip requirements -------------------------------------------------
# tb-nightly
# future

# Equivalent conda commands ----------------------------------------------------
# conda update -n base -c defaults conda
# conda install -yc anaconda future numpy opencv matplotlib tqdm pillow
# conda install -yc conda-forge scikit-image pycocotools tensorboard
# conda install -yc spyder-ide spyder-line-profiler
# conda install -yc pytorch pytorch torchvision

torchvision~=0.8.1
django-pydash-app~=1.4.4